abstract rl agent
